+++ delivery/src/main/java/com/nowayback/delivery/application/delivery/service/HubClient.java
@@ -1,4 +1,4 @@
-package com.nowayback.delivery.application.service;
+package com.nowayback.delivery.application.delivery.service;
 
 import java.util.UUID;
