+++ delivery/src/main/java/com/nowayback/delivery/application/deliveryroute/command/UpdateDeliveryRouteInfoCommand.java
@@ -1,4 +1,4 @@
-package com.nowayback.delivery.application.command;
+package com.nowayback.delivery.application.deliveryroute.command;
 
 public record UpdateDeliveryRouteInfoCommand(
         int actualDistanceMeters,
