+++ delivery/src/main/java/com/nowayback/delivery/application/deliveryroute/exception/DeliveryRouteApplicationException.java
@@ -1,4 +1,4 @@
-package com.nowayback.delivery.application.exception;
+package com.nowayback.delivery.application.deliveryroute.exception;
 
 import com.nowayback.common.exception.GlobalException;
