+++ delivery/src/test/java/com/nowayback/delivery/fixture/DeliveryManagerFixture.java
@@ -2,7 +2,7 @@ package com.nowayback.delivery.fixture;
 
 import com.nowayback.delivery.application.deliverymanager.command.CreateDeliveryManagerCommand;
 import com.nowayback.delivery.application.deliverymanager.dto.DeliveryManagerResult;
-import com.nowayback.delivery.application.service.UserClient;
+import com.nowayback.delivery.application.deliverymanager.service.UserClient;
 import com.nowayback.delivery.domain.deliverymanager.entity.DeliveryManager;
 import com.nowayback.delivery.domain.deliverymanager.vo.DeliveryManagerType;
 import com.nowayback.delivery.domain.deliverymanager.vo.DeliverySequence;
